requirements.txt
streamlit
google-genai
deepgram-sdk
python-dotenv
streamlit-webrtc